Refactoring epoch configs
